+++ Example/GBHFacebookImagePicker/ViewController.swift
@@ -19,7 +19,7 @@ class ViewController: UIViewController {
 
     // MARK: - Var 
 
-    fileprivate var imageModels = [GBHFacebookImage]() {
+    fileprivate var imageModels = [FacebookImage]() {
         didSet {
             DispatchQueue.main.async {
                 self.tableView.reloadData()
@@ -61,58 +61,58 @@ class ViewController: UIViewController {
 
     fileprivate func someCustomisation() {
         // Navigation bar title 
-        GBHFacebookImagePicker.pickerConfig.textConfig.localizedTitle = "MyCustomTitle"
+        FacebookImagePicker.pickerConfig.textConfig.localizedTitle = "MyCustomTitle"
 
         // Navigation barTintColor
-        GBHFacebookImagePicker.pickerConfig.uiConfig.navBarTintColor = UIColor.red
+        FacebookImagePicker.pickerConfig.uiConfig.navBarTintColor = UIColor.red
 
         // Close button color 
-        GBHFacebookImagePicker.pickerConfig.uiConfig.closeButtonColor = UIColor.white
+        FacebookImagePicker.pickerConfig.uiConfig.closeButtonColor = UIColor.white
 
         // Global backgroundColor 
-        GBHFacebookImagePicker.pickerConfig.uiConfig.backgroundColor = UIColor.red
+        FacebookImagePicker.pickerConfig.uiConfig.backgroundColor = UIColor.red
 
         // Navigation bar title color
-        GBHFacebookImagePicker.pickerConfig.uiConfig.navTitleColor = UIColor.white
+        FacebookImagePicker.pickerConfig.uiConfig.navTitleColor = UIColor.white
 
         // Navigation bar tintColor
-        GBHFacebookImagePicker.pickerConfig.uiConfig.navTintColor = UIColor.white
+        FacebookImagePicker.pickerConfig.uiConfig.navTintColor = UIColor.white
 
         // Album's name color 
-        GBHFacebookImagePicker.pickerConfig.uiConfig.albumsTitleColor = UIColor.white
+        FacebookImagePicker.pickerConfig.uiConfig.albumsTitleColor = UIColor.white
 
         // Album's count color 
-        GBHFacebookImagePicker.pickerConfig.uiConfig.albumsCountColor = UIColor.white
+        FacebookImagePicker.pickerConfig.uiConfig.albumsCountColor = UIColor.white
 
         // Maximum selected pictures 
-        GBHFacebookImagePicker.pickerConfig.maximumSelectedPictures = 4
+        FacebookImagePicker.pickerConfig.maximumSelectedPictures = 4
 
         // Display tagged album 
-        GBHFacebookImagePicker.pickerConfig.textConfig.localizedTaggedAlbumName = "Tagged photos"
+        FacebookImagePicker.pickerConfig.textConfig.localizedTaggedAlbumName = "Tagged photos"
 
         // Tagged album name
-        GBHFacebookImagePicker.pickerConfig.displayTaggedAlbum = true
+        FacebookImagePicker.pickerConfig.displayTaggedAlbum = true
 
         // Number of picture per row (4 by default)
-        GBHFacebookImagePicker.pickerConfig.picturePerRow = 3
+        FacebookImagePicker.pickerConfig.picturePerRow = 3
 
         // Space beetween album photo cell (1.5 by default)
-        GBHFacebookImagePicker.pickerConfig.cellSpacing = 2.0
+        FacebookImagePicker.pickerConfig.cellSpacing = 2.0
 
         // Perform animation on picture tap (true by default)
-        GBHFacebookImagePicker.pickerConfig.performTapAnimation = true
+        FacebookImagePicker.pickerConfig.performTapAnimation = true
 
         // Show check style with image and layer (true by default)
-        GBHFacebookImagePicker.pickerConfig.uiConfig.showCheckView = true
+        FacebookImagePicker.pickerConfig.uiConfig.showCheckView = true
 
         // Change checkview background color
-        GBHFacebookImagePicker.pickerConfig.uiConfig.checkViewBackgroundColor = UIColor.red
+        FacebookImagePicker.pickerConfig.uiConfig.checkViewBackgroundColor = UIColor.red
 
         // Preview photos size (normal by default)
-        GBHFacebookImagePicker.pickerConfig.uiConfig.previewPhotoSize = .full
+        FacebookImagePicker.pickerConfig.uiConfig.previewPhotoSize = .full
 
         // Show the "Select all" button 
-        GBHFacebookImagePicker.pickerConfig.allowAllSelection = true
+        FacebookImagePicker.pickerConfig.allowAllSelection = true
     }
 
     // MARK: - Action
@@ -121,14 +121,14 @@ class ViewController: UIViewController {
         print(self, #function)
 
         // Init picker 
-        let picker = GBHFacebookImagePicker()
+        let picker = FacebookImagePicker()
 
         // Allow multiple selection (false by default)
-        GBHFacebookImagePicker.pickerConfig.allowMultipleSelection = true
-        GBHFacebookImagePicker.pickerConfig.uiConfig.previewPhotoSize = .full
-        GBHFacebookImagePicker.pickerConfig.allowAllSelection = true
-        GBHFacebookImagePicker.pickerConfig.picturePerRow = 3
-        GBHFacebookImagePicker.pickerConfig.displayTaggedAlbum = true
+        FacebookImagePicker.pickerConfig.allowMultipleSelection = true
+        FacebookImagePicker.pickerConfig.uiConfig.previewPhotoSize = .full
+        FacebookImagePicker.pickerConfig.allowAllSelection = true
+        FacebookImagePicker.pickerConfig.picturePerRow = 3
+        FacebookImagePicker.pickerConfig.displayTaggedAlbum = true
 
         // Make some customisation
         //self.someCustomisation()
@@ -140,17 +140,17 @@ class ViewController: UIViewController {
 
     @IBAction func doDeleteClick(_ sender: Any) {
         // Clear data src 
-        self.imageModels = [GBHFacebookImage]()
+        self.imageModels = [FacebookImage]()
     }
 }
 
-extension ViewController: GBHFacebookImagePickerDelegate {
+extension ViewController: FacebookImagePickerDelegate {
 
     // MARK: - GBHFacebookImagePicker Protocol
 
     func facebookImagePicker(imagePicker: UIViewController,
-                             successImageModels: [GBHFacebookImage],
-                             errorImageModels: [GBHFacebookImage],
+                             successImageModels: [FacebookImage],
+                             errorImageModels: [FacebookImage],
                              errors: [Error?]) {
         // Append selected image(s)
         // Do what you want with selected image