+++ GBHFacebookImagePicker/Classes/Controller/Helpers/AssetsController.swift
@@ -5,7 +5,7 @@
 //  Created by Florian Gabach on 02/10/2016.
 //  Copyright (c) 2016 Florian Gabach <contact@floriangabach.fr>
 
-public class AssetsController {
+final class AssetsController {
 
     /// Get image from bundle
     ///