+++ GBHFacebookImagePicker/Classes/Controller/Public-FacebookImagePicker/FacebookImagePicker.swift
@@ -7,10 +7,10 @@
 
 import UIKit
 
-public class GBHFacebookImagePicker: NSObject {
+public class FacebookImagePicker: NSObject {
 
     // Picker configuration settings
-    open static var pickerConfig = GBHFacebookPickerConfig()
+    open static var pickerConfig = FacebookPickerConfig()
 
     /// Present album picker
     ///
@@ -18,14 +18,14 @@ public class GBHFacebookImagePicker: NSObject {
     ///   - from: controller where we want present the picker
     ///   - delegate: delegate for GBHFacebookImagePickerDelegate
     public final func presentFacebookAlbumImagePicker(from: UIViewController,
-                                                      delegate: GBHFacebookImagePickerDelegate) {
+                                                      delegate: FacebookImagePickerDelegate) {
 
         // Create album picker
         let albumPicker = FacebookAlbumController()
         albumPicker.delegate = delegate
 
         // Embed in navigation controller
-        let navigationController = GBHFacebookNavigationController(rootViewController: albumPicker)
+        let navigationController = FacebookImagePickerNavigationController(rootViewController: albumPicker)
 
         // Present the picker 
         from.present(navigationController, animated: true)