+++ GBHFacebookImagePicker/Classes/Model/AssetImage.swift
@@ -1,5 +1,5 @@
 //
-//  GBHAssetImage.swift
+//  AssetImage.swift
 //  Bolts
 //
 //  Created by Florian Gabach on 10/09/2017.
@@ -8,7 +8,7 @@
 import Foundation
 
 /// Struct to define assets name 
-struct GBHAssetImage {
+struct AssetImage {
     /// Default image loader name
     static let loader = "GBHFacebookImagePickerDefaultImageLoading"