+++ GBHFacebookImagePicker/Classes/View/SelectedView.swift
@@ -1,5 +1,5 @@
 //
-//  GBHSelectedView.swift
+//  SelectedView.swift
 //  Pods
 //
 //  Created by Florian Gabach on 03/05/2017.
@@ -15,7 +15,7 @@ public enum CheckViewPosition {
     case bottomRight
 }
 
-final class GBHSelectedView: UIView {
+final class SelectedView: UIView {
 
     // MARK: - Private var
 
@@ -29,7 +29,7 @@ final class GBHSelectedView: UIView {
         super.init(frame: frame)
 
         // Prepare
-        if  GBHFacebookImagePicker.pickerConfig.uiConfig.showCheckView {
+        if  FacebookImagePicker.pickerConfig.uiConfig.showCheckView {
             self.prepareView()
         }
     }
@@ -38,7 +38,7 @@ final class GBHSelectedView: UIView {
         super.init(coder: aDecoder)
 
         // Prepare
-        if GBHFacebookImagePicker.pickerConfig.uiConfig.showCheckView {
+        if FacebookImagePicker.pickerConfig.uiConfig.showCheckView {
             self.prepareView()
         }
     }
@@ -56,7 +56,7 @@ final class GBHSelectedView: UIView {
                                         height: self.checkMarkViewSize.height)
 
         // Checkmark background color
-        let customColor = GBHFacebookImagePicker.pickerConfig.uiConfig.checkViewBackgroundColor
+        let customColor = FacebookImagePicker.pickerConfig.uiConfig.checkViewBackgroundColor
         let backgroundColor = customColor ?? Colors.facebookColor
 
         self.checkMarkView?.layer.borderWidth = 1.5
@@ -75,8 +75,8 @@ final class GBHSelectedView: UIView {
         self.prepareConstraint()
 
         // Border
-        self.layer.borderColor = GBHFacebookImagePicker.pickerConfig.uiConfig.selectedBorderColor?.cgColor
-        self.layer.borderWidth = GBHFacebookImagePicker.pickerConfig.uiConfig.selectedBorderWidth
+        self.layer.borderColor = FacebookImagePicker.pickerConfig.uiConfig.selectedBorderColor?.cgColor
+        self.layer.borderWidth = FacebookImagePicker.pickerConfig.uiConfig.selectedBorderWidth
     }
 
     fileprivate func prepareCheckMarkView() {
@@ -84,7 +84,7 @@ final class GBHSelectedView: UIView {
 
         // Add checkmark image from assets
         let checkImageView = UIImageView()
-        checkImageView.image = AssetsController.getImage(name: GBHAssetImage.checkMark)
+        checkImageView.image = AssetsController.getImage(name: AssetImage.checkMark)
         checkImageView.backgroundColor = .clear
         checkView.addSubview(checkImageView)
 
@@ -132,7 +132,7 @@ final class GBHSelectedView: UIView {
         var margin: CGFloat = 2
 
         // Extend margin if borderWidth is chosen
-        margin += GBHFacebookImagePicker.pickerConfig.uiConfig.selectedBorderWidth
+        margin += FacebookImagePicker.pickerConfig.uiConfig.selectedBorderWidth
 
         // Place vertical constraint
         self.applyVerticalConstraints(margin: margin)
@@ -163,7 +163,7 @@ final class GBHSelectedView: UIView {
 
     fileprivate func applyVerticalConstraints(margin: CGFloat) {
         guard let checkMarkView = self.checkMarkView else { return }
-        switch GBHFacebookImagePicker.pickerConfig.uiConfig.placeCheckView {
+        switch FacebookImagePicker.pickerConfig.uiConfig.placeCheckView {
         case .topLeft, .topRight:
             self.addConstraint(NSLayoutConstraint(item: checkMarkView,
                                                   attribute: NSLayoutConstraint.Attribute.top,
@@ -185,7 +185,7 @@ final class GBHSelectedView: UIView {
 
     fileprivate func applyHorizontalConstraints(margin: CGFloat) {
         guard let checkMarkView = self.checkMarkView else { return }
-        switch GBHFacebookImagePicker.pickerConfig.uiConfig.placeCheckView {
+        switch FacebookImagePicker.pickerConfig.uiConfig.placeCheckView {
         case .topLeft, .bottomLeft:
             self.addConstraint(NSLayoutConstraint(item: checkMarkView,
                                                   attribute: NSLayoutConstraint.Attribute.leading,